Fix health check enable/disable problem
The health manager is treating all health policies as always enabled,
this is wrong. When users want to disable a health policy, the health
manager should be notified. Such a status change need to be persisted
into database so that even an engine dies, other engines will know how
to take an appropriate action.

Change-Id: I2e564ac78310471c25b5fe2a976f551424644126
Closes-Bug: #1640994